betting_futbol/requirements.txt
# =====================================
# CORE FRAMEWORK
# =====================================

Flask==2.3.3
Werkzeug==2.3.7

# =====================================
# DATABASE
# =====================================

Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5

# =====================================
# SECURITY
# =====================================

Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0

# =====================================
# API & REQUESTS
# =====================================

requests==2.31.0
python-dotenv==1.0.0

# =====================================
# ADMIN PANEL & FORMS
# =====================================

Flask-WTF==1.2.1
WTForms==3.1.2

# =====================================
# PERFORMANCE
# =====================================

Flask-Caching==2.1.0
redis==5.0.1

# =====================================
# REAL TIME (LIVE BETTING)
# =====================================

Flask-SocketIO==5.3.6
eventlet==0.33.3

# =====================================
# PRODUCTION SERVER
# =====================================

gunicorn==21.2.0

# =====================================
# UTILITIES
# =====================================

python-dateutil==2.8.2
pytz==2023.3

# =====================================
# LOGGING & MONITORING
# =====================================

loguru==0.7.2

# =====================================
# TELEGRAM BOT
# =====================================

python-telegram-bot==20.7